+++ plugins/postcss-rebase-url/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Rebase URL
 
+### Unreleased (major)
+
+- Updated: Support for Node v18+ (major).
+
 ### 1.0.7
 
 _March 13, 2024_